+++ src/api.js
@@ -1,8 +1,5 @@
 import { ApiCaller } from '@tokend/js-sdk'
 
-const HORIZON_VERSION_PREFIX = 'v3'
-const IDENTITIES_PATH = 'identities'
-
 let _api = null
 
 export class Api {
@@ -39,12 +36,7 @@ export class Api {
    * @returns {String} - endpoint path, containing horizon version
    */
   static _getEndpoint (path) {
-    // TODO: refactor it. maked near deadline
-    if (path.split('/')[0] !== IDENTITIES_PATH) {
-      return `/${HORIZON_VERSION_PREFIX}/${path}`
-    } else {
-      return `/${path}`
-    }
+    return `/${path}`
   }
 
   /**